computer_basics/lessons/api.py
from lessons.models import Lesson
from rest_framework import viewsets, permissions
from .serializers import LessonSerializer

class LessonViewSet(viewsets.ModelViewSet):
    queryset = Lesson.objects.all()
    permission_classes = [
        permissions.AllowAny
    ]
    serializer_class = LessonSerializer