requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
cffi==1.14.4
colorama==0.4.4
cryptography==3.3.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.4
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.12.2
filelock==3.0.12
gunicorn==20.0.4
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.1.0
psycopg2==2.8.6
pycparser==2.20
pylint==2.6.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.3.1
virtualenv-clone==0.5.4
whitenoise==5.2.0
wrapt==1.12.1
